refactor(indexer): restructure restore module
